Netcore extensions rework in progress
