+++ .editorconfig
@@ -232,7 +232,7 @@ dotnet_diagnostic.CA1001.severity = none
 dotnet_diagnostic.CA1031.severity = none
 
 # Do not nest type
-dotnet_diagnostic.CA1034.severity = none
+# dotnet_diagnostic.CA1034.severity = none
 
 # Do not declare visible instance fields
 dotnet_diagnostic.CA1051.severity = none
